change carouseView
挑战首页基本完成

+++ XFitClub/AppDelegate.m
@@ -87,7 +87,7 @@
 
 //计时
 -(void)countAction{
-    NSLog(@"%li",count++);
+    NSLog(@"%li",(long)count++);
 }
 //申请后台
 -(void)beginTask
